PopupMenu List Investigation
Bunch of changes to see if I could get the PopupMenu to be rebuilt with our current SwiftUI List. Short answer is no.
Here are some of the issues I was running into:
The List wants to create the cells itself, while the PopupMenu wanted to accept them from somewhere else.
The ListCells aren't built to have spacing for a missing icon (I think, could be wrong)
The ListCells can't recolor the leading image, since we store it as just a UIView.
The ListCells aren't able to be disabled, or set to be selected when they're made. I think at least the selection issue could be fixed while we're changing the way that cell creation is handled. Though, for sure the current list doesn't know anything about selecting multiple cells at the same time, so we might want some sort of check about that.
For some reason the List wasn't playing well with the safe areas of the Drawer controller.
The list sometimes animates, probably just need to add an option to have the animation not be set to default, but I didn't really look into that one.
There's a weird gap in the stack between the PopupMenu header and the rest of the list. Not sure what's going on there.